Refaktor: InMemory- és JpaCartService közös részek kiemelése
